Rename `trace_data_utils` -> `trace_data`
