+++ crates/cairo-annotations/tests/integration/mod.rs
@@ -1,4 +1,4 @@
 mod coverage_annotations;
 mod map_pcs_to_sierra_statement_ids;
 mod profiler_annotations;
-mod trace_data_utils;
+mod trace_data;
