+++ crates/cairo-annotations/tests/integration/trace_data.rs
